fix: tests + bugs
