+++ packages/wallet-lib/fixtures/wallets/2a331817b9d6bf85100ef0/getFixtureAccountWithStorage.js
@@ -1,3 +1,4 @@
+const _ = require('lodash')
 const walletStoreMock = require('./wallet-store.json');
 const chainStoreMock = require('./chain-store.json');
 const Storage = require('../../../src/types/Storage/Storage');
@@ -8,7 +9,8 @@ const generateNewPaths = require("../../../src/types/Account/methods/generateNew
 const addDefaultPaths = require("../../../src/types/Account/methods/addDefaultPaths");
 
 module.exports = (opts = {}) => {
-    const { walletId } = walletStoreMock;
+  const _walletStoreMock = _.cloneDeep(walletStoreMock)
+  const { walletId } = _walletStoreMock;
 
     const mockedAccount = {
         walletId,
@@ -27,7 +29,7 @@ module.exports = (opts = {}) => {
     mockedAccount.storage.createChainStore('testnet');
 
     const walletStore = mockedAccount.storage.getWalletStore(walletId);
-    walletStore.importState(walletStoreMock);
+    walletStore.importState(_walletStoreMock);
     walletStore.createPathState(mockedAccount.accountPath);
 
     mockedAccount.storage.getChainStore('testnet').importState(chainStoreMock);
