+++ packages/wallet-lib/fixtures/wallets/2a331817b9d6bf85100ef0/wallet-store.json
@@ -1,15 +1,6 @@
 {
   "walletId": "6101b44d50",
-  "state": {
-    "mnemonic": null,
-    "paths": {
-      "m/0": {
-        "path": "m/0",
-        "addresses": {
-          "0": "ycDeuTfs4U77bTb5cq17dame28zdWHVYfk"
-        }
-      }
-    },
-    "identities": {}
+  "lastKnownBlock": {
+    "height": 11703
   }
 }
