+++ packages/wallet-lib/fixtures/wallets/apart-trip-dignity/getFixtureAccountWithStorage.js
@@ -1,3 +1,4 @@
+const _ = require('lodash');
 const walletStoreMock = require('./wallet-store.json');
 const chainStoreMock = require('./chain-store.json');
 const Storage = require('../../../src/types/Storage/Storage');
@@ -8,7 +9,10 @@ const generateNewPaths = require("../../../src/types/Account/methods/generateNew
 const addDefaultPaths = require("../../../src/types/Account/methods/addDefaultPaths");
 
 module.exports = (opts = {}) => {
-    const { walletId } = walletStoreMock;
+    const _walletStoreMock = _.cloneDeep(walletStoreMock)
+    const _chainStoreMock = _.cloneDeep(chainStoreMock)
+
+    const { walletId } = _walletStoreMock;
 
     const mockedWallet = {
       walletId,
@@ -25,9 +29,9 @@ module.exports = (opts = {}) => {
     }), { isMasterKeyChain: true });
 
     const walletStore = mockedWallet.storage.getWalletStore(walletId);
-    walletStore.importState(walletStoreMock);
+    walletStore.importState(_walletStoreMock);
     const chainStore = mockedWallet.storage.getChainStore('testnet');
-    chainStore.importState(chainStoreMock);
+    chainStore.importState(_chainStoreMock);
 
     const mockedAccount0 = {
         walletId,
