+++ packages/wallet-lib/fixtures/wallets/apart-trip-dignity/wallet-store.json
@@ -1,116 +1,6 @@
 {
   "walletId": "d6143ef4e6",
-  "state": {
-    "mnemonic": null,
-    "paths": {
-      "m/44'/1'/1'": {
-        "path": "m/44'/1'/1'",
-        "addresses": {
-          "m/0/0": "yYJmzWey5kNecAThet5BFxAga1F4b4DKQ2",
-          "m/0/1": "yNCqctyQaq51WU1hN5aNwsgMsZ5fRiB7GY",
-          "m/0/2": "yNPbYz5cZKw2EwxtkL3VSVzPi2FYp9VKjQ",
-          "m/0/3": "ybsGWzsnSCAZufgSeUjScVxqEdved99UM2",
-          "m/0/4": "yfNHuPojk8XKWP5nuueDptX4nM7qToudgx",
-          "m/0/5": "yXxLnDkk6s8h1PSnYaFM6MAyRarc1Kc1rY",
-          "m/0/6": "yipResSzN2zUvL7UYkmptKKmQTv7sNssRn",
-          "m/0/7": "yZPtNwimHdRiKYbNQW49qezw1Kc1YwUJeT",
-          "m/0/8": "yPMjYYfQbga2nBiuqqfUyX41U1vwRZ8fG8",
-          "m/0/9": "yLueLWWcLQsaXQ8D5o9tcyo8tfTxMWXvG4",
-          "m/0/10": "yN8gzgsc1RVjXThMQT5qZH2jjpnMymz6zP",
-          "m/0/11": "yPQLWBNwMdLxUW2oUwHGwQtfyYxD41BARJ",
-          "m/0/12": "yg5g2AfWFdwWexWGfbSXYbUHf1y5WWrFPs",
-          "m/0/13": "yWyABu4naV1Jzw7w9sn1gqhebPRSkCndsS",
-          "m/0/14": "ycuUPzUBjhKyUjezQR1LNot79a6C4aRLaR",
-          "m/0/15": "yQ7YjvAXgDAUCekveHVjr6NBveXrUemVno",
-          "m/0/16": "yi8bghcw627cMGpuH4bJqH6bqR5ywv1NLH",
-          "m/0/17": "yizHu8i2rfwzwBgnJ62s2WUe6wLoDjne6N",
-          "m/0/18": "yW1u3tySeUKAKJsz7sjZFyjUiTyKLB6xBv",
-          "m/0/19": "yNaSkdy1Q8JNubUdbLMGsGf7sTRofEJYZq",
-          "m/1/0": "yXMrw79LPgu78EJsfGGYpm6fXKc1EMnQ49",
-          "m/1/1": "yh6Hcyipdvp6WJpQxjNbaXP4kzPQUJpY3n",
-          "m/1/2": "yNphpXuaTZRpU9FBh2W7NkUYcr3kBDE8me",
-          "m/1/3": "yXFppDT59xYD41mT2pmAdnvr7aZEFdgdrN",
-          "m/1/4": "yeKGAiiEHBGRujvLoYewA77jDDpeDamxvF",
-          "m/1/5": "yaxTG66CVzKgHhHZXojRHC9ztLTvz3fwdT",
-          "m/1/6": "yYw6qU7dwGoELZkSTj3oSKRpM4U8qTMc1U",
-          "m/1/7": "yQE2MksEnSfbeNre19oja9Jj8tvpj64C5a",
-          "m/1/8": "yaRnvHo8oLvVmv46vMj5XPbDJouQSnmcLT",
-          "m/1/9": "yj5ofWf2uYQQkSavYm2WXgu1QkaZCyP3Cm",
-          "m/1/10": "yUCjGmEwrHJwNDrE1o2rMre6MkSbiE6yz7",
-          "m/1/11": "yfJzd1nE2rEqz5XEurD6vs4ykizwmw9xTv",
-          "m/1/12": "yUk8U3jRZMHKVTa1eFDEtZpa1G4E13FP4d",
-          "m/1/13": "yMr59YWQFCADq4FbWrtxDUtMwwshSrmAyK",
-          "m/1/14": "yetSehBupzGS9yps5ogqARUGmTMAs2xVcQ",
-          "m/1/15": "yNcESKLwriNrhM6EyoSpZEXrzdY3uht92T",
-          "m/1/16": "yN2FihGU7KdaEspp39bKrhsHypeyeYzoM2",
-          "m/1/17": "yirpWLxHuhwFzA6LfUPKUh1Ke9RB9BUjit",
-          "m/1/18": "yVDN66vvdshWdNzhUaQNB6xExAHkzs1zj8",
-          "m/1/19": "yPzofnEhRVfDisL2nCUJtAoSHkuyMirHZS",
-          "m/0/20": "yjG4jiCMNbx3MyFCAsaNaU29CBVmgPf8hS",
-          "m/0/21": "yWrEYTtydaZCyJoKE1vzX4e4RXDNPGdnw9",
-          "m/1/20": "yM7Smr8HEhbRW4CXga898brATrJSsk1QUh",
-          "m/1/21": "yUprvaE5KrKVPgABmGtxba9MmZqDzKQHom"
-        }
-      },
-      "m/44'/1'/0'": {
-        "path": "m/44'/1'/0'",
-        "addresses": {
-          "m/0/0": "yTwEca67QSkZ6axGdpNFzWPaCj8zqYybY7",
-          "m/0/1": "yercyhdN9oEkZcB9BsW5ktFaDxFEuK6qXN",
-          "m/0/2": "ygk3GCSba2J3L9G665Snozhj9HSkh5ByVE",
-          "m/0/3": "ybuL6rM6dgrKzCg8s99f3jxGuv5oz5JcDA",
-          "m/0/4": "ygHAVkMtYSqoTWHebDv7qkhMV6dHyuRsp2",
-          "m/0/5": "yMLhEsiP2ajSh8STmXnNmkWXtoHsmawZxd",
-          "m/0/6": "yj8rRKATAUHcAgXvNZekob58xKm2oNyvhv",
-          "m/0/7": "yhaAB6e8m3F8zmGX7WAVYa6eEfmSrrnY8x",
-          "m/0/8": "yiXh4Yo5djG6QH8WzXkKm5EFzqLRJWakXz",
-          "m/0/9": "yQYv3Um6DsdtANo1ZPTUte75wAGMstLRex",
-          "m/0/10": "yiYPJmu7eEm1cXUNumQRdjv1fvPhsfgMS4",
-          "m/0/11": "yii4aUZhNfL6EWN9KAgAFrJzGJmqHnF4wx",
-          "m/0/12": "yLpTquSct2SGz2Ka45uTPDd81Kzro2Jt2k",
-          "m/0/13": "yMiJtpzb1Qthy9TGnavsf5NZ6EZZa4j9q3",
-          "m/0/14": "yacgSfW7RkwWakEZPg8USAVdzCypiG3vxS",
-          "m/0/15": "yVvrmoRPFLy6nUpCQBT8ZExxF5wF3DhiGU",
-          "m/0/16": "yaJf2aG6cFUtfv4o6TuEKsh5kr4xq5iAY4",
-          "m/0/17": "yfardJQ4ucgWLKQPaRHGMRMbSGm5H4ExJR",
-          "m/0/18": "yLSCqx7dcM5JKR2fG7vHbF2axMvuYqomaw",
-          "m/0/19": "yVij8XpJ78LM5hepSV1KF7T8vRpUEXCpK5",
-          "m/1/0": "yNDpPsJqXKM36zHSNEW7c1zSvNnrZ699FY",
-          "m/1/1": "yLk4Hw3w4zDudrDVP6W8J9TggkY57zQUki",
-          "m/1/2": "yirJaK8KCE5YAmwvLadizqFw3TCXqBuZXL",
-          "m/1/3": "yhdRfg5gNr587dtEC4YYMcSHmLVEGqqtHc",
-          "m/1/4": "yYwKP1FQae5kbjXkmuirGx6Xzf8NzHpLqW",
-          "m/1/5": "yX9gmsm8aSxZZjYhq4w35aidT7qbhcpNjU",
-          "m/1/6": "ybgXCTGMHEBbQeUib8c3xAjtGAc12XtWiU",
-          "m/1/7": "yS31WpdMT2b34uL9C37fbUoACHhiupHCyP",
-          "m/1/8": "yTSpFqRoX3vyN286AUtKKhgmX5Xb41YKQe",
-          "m/1/9": "yQU5YsqN7psTTASuYbcMi7N5nNZGaxXb2X",
-          "m/1/10": "yVGGFj9BLgEab5rucSGLC6UGVLQKB4U1wJ",
-          "m/1/11": "yQCh5yYCHEbJzgSJE9rdHiqXHidKm3kwr5",
-          "m/1/12": "yX7T3Ac3yaLk5CTC5UaR93Fc7SjYkeT5hn",
-          "m/1/13": "yXx3WXq8kYNPbYEg5U6bL8Xfih4g5LCYVo",
-          "m/1/14": "yYnLMTz3jCi2KKKNuo3TVkEAGyUFg8tgkJ",
-          "m/1/15": "yiKa1dA6B4tSTNJqJP9Y5pQfQEffnQQDTL",
-          "m/1/16": "yf7vcuDnE9DVhXdMfBMQQTEi43otYQzkWE",
-          "m/1/17": "yTmSmocwERCeRHqNNG5SbpYKUra1HTmj8m",
-          "m/1/18": "yivUe5NeJsGsREwPQZUGYaTSwWB3E1oLcz",
-          "m/1/19": "ygfsZojdfW9UjCRU4ra95Aq6YgCC7UqZFx",
-          "m/1/20": "yU9fdXaUVtefwDZvxjJAr9xj1z2MtYi34A",
-          "m/1/21": "yXgMN6FgrgZCnTN1vhoZMh8afKMBmi3JC4",
-          "m/1/22": "yiqaCbXscvR8y3VFYMzdaKCaAGuDuZxMzt",
-          "m/0/20": "ydJpjuJGossAZR7S5oS7cWvjygEwoj8Xwp",
-          "m/0/21": "yW3TmWnmhvpxRbgFcQ8oXqDRkn3RhRH6jj",
-          "m/0/22": "yRegVX85DThKRkH8C61TtRacfzrkiBfNy5",
-          "m/0/23": "yPtDCqDFRe1JuDp8pvdiEMQMz2erGwS3VG",
-          "m/0/24": "yM9pSw3L4oBfG7uQL5o522Hu3WTvy9awgZ",
-          "m/0/25": "yNC6qYJYungzuk5XUynDFKCn54Dy8ngox4",
-          "m/0/26": "yR5KcLr1bceLT4teTk2qoJx6pFLik1zyzL",
-          "m/0/27": "yRrKLGJa9JmdjBWvrHtedKjHTao6CRDTKf",
-          "m/0/28": "yP5dShZBydpbEzgGoXL6kcjv2KzervRrYB",
-          "m/1/23": "ydcgWDxheSxrLAqDBP4JXBndMCzUNf77gq"
-        }
-      }
-    },
-    "identities": {}
+  "lastKnownBlock": {
+    "height": 11703
   }
 }
